requirements.txt
alembic==1.14.0
Flask==3.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
python-dotenv==1.0.1
prometheus-flask-exporter==0.23.1